feat: updated license in create model

+++ budapp/model_ops/services.py
@@ -17,9 +17,10 @@
 """The model ops services. Contains business logic for model ops."""
 
 import os
+import re
 import tempfile
 from typing import Any, Dict, List, Literal, Optional, Tuple
-from urllib.parse import urlparse
+from urllib.parse import unquote, urlparse
 from uuid import UUID, uuid4
 
 import aiohttp
@@ -50,14 +51,15 @@ from ..commons.constants import (
     BENCHMARK_FIELDS_LABEL_MAPPER,
     BENCHMARK_FIELDS_TYPE_MAPPER,
     BUD_INTERNAL_WORKFLOW,
+    COMMON_LICENSE_MINIO_OBJECT_NAME,
     HF_AUTHORS_DIR,
+    MINIO_LICENSE_OBJECT_NAME,
     BaseModelRelationEnum,
     BudServeWorkflowStepEventName,
     CloudModelStatusEnum,
     ClusterStatusEnum,
     CredentialTypeEnum,
     EndpointStatusEnum,
-    ModelLicenseObjectTypeEnum,
     ModelProviderTypeEnum,
     ModelSecurityScanStatusEnum,
     ModelSourceEnum,
@@ -1376,57 +1378,18 @@ class LocalModelWorkflowService(SessionMixin):
                     }
                 )
 
-        license_object_type, updated_license_url = await self._license_url_type_identifier(
-            local_path, license_url, model_id
-        )
-
         license_data = ModelLicensesCreate(
             name=license_name,
-            url=updated_license_url,
+            url=license_url,
             faqs=updated_license_faqs if updated_license_faqs else None,
             model_id=model_id,
             license_type=license_type,
             description=license_description,
             suitability=license_suitability,
-            data_type=license_object_type,
         )
         return await ModelLicensesDataManager(self.session).insert_one(
             ModelLicenses(**license_data.model_dump(exclude_none=True))
         )
-
-    async def _license_url_type_identifier(
-        self, local_path: str, license_url: str, model_id: UUID
-    ) -> Tuple[ModelLicenseObjectTypeEnum, str]:
-        """Identify license url type.
-
-        Args:
-            local_path (str): Local path.
-            license_url (str): License url.
-
-        Returns:
-            str: Model license object type.
-        """
-        if license_url.startswith(local_path):
-            # Check if object exists in minio
-            model_store = ModelStore()
-            is_minio_object_exists = model_store.check_file_exists(app_settings.minio_bucket, license_url)
-            if is_minio_object_exists:
-                # Download to a temp file and upload to minio
-                with tempfile.TemporaryDirectory() as temp_dir:
-                    minio_store = ModelStore()
-                    file_name = os.path.basename(license_url)
-                    download_path = os.path.join(temp_dir, file_name)
-                    minio_store.download_object(app_settings.minio_bucket, license_url, download_path)
-
-                    # License object name
-                    license_object_name = f"{model_id}/{file_name}"
-                    minio_store.upload_file(app_settings.minio_model_bucket, download_path, license_object_name)
-                    return ModelLicenseObjectTypeEnum.MINIO, license_object_name
-            else:
-                logger.error("License url %s does not exist in minio.", license_url)
-                return ModelLicenseObjectTypeEnum.URL, license_url
-        else:
-            return ModelLicenseObjectTypeEnum.URL, license_url
 
     @staticmethod
     async def get_base_model_relation(model_tree: dict) -> Optional[BaseModelRelationEnum]:
@@ -2082,20 +2045,33 @@ class ModelService(SessionMixin):
             file = data.pop("license_file")
 
             # Save to minio
-            license_object_name = await self._save_license_to_minio(file, model_id)
+            license_object_name = await self._save_license_file_to_minio(file, model_id)
 
             # Create or update license entry
-            await self._create_or_update_license_entry(
-                model_id, file.filename, license_object_name, current_user_id, ModelLicenseObjectTypeEnum.MINIO
-            )
+            await self._create_or_update_license_entry(model_id, file.filename, license_object_name, current_user_id)
 
         elif data.get("license_url"):
             # If a license URL is provided, store the URL in the DB instead of the file path
             license_url = data.pop("license_url")
-            filename = license_url.split("/")[-1]  # Extract filename from the URL
-            filename = filename if filename else "sample license"
+
+            # get file name with extension from url
+            parsed = urlparse(license_url)
+            path = unquote(parsed.path)
+            filename = path.split("/")[-1]
+
+            # Handle query parameters
+            if "?" in filename:
+                filename = filename.split("?")[0]
+
+            # Validate filename
+            if not filename or not re.match(r"^[\w\-\.]+$", filename):
+                filename = "LICENSE"
+
+            # Save to minio
+            license_object_name = await self._save_license_url_to_minio(license_url, filename, model_id)
+
             await self._create_or_update_license_entry(
-                model_id, filename, str(license_url), current_user_id, ModelLicenseObjectTypeEnum.URL
+                model_id, filename, license_object_name, current_user_id
             )  # TODO: modify filename arg when license service implemented
 
         # Add papers if provided
@@ -2106,11 +2082,11 @@ class ModelService(SessionMixin):
         # Update model with validated data
         await ModelDataManager(self.session).update_by_fields(db_model, data)
 
-    async def _save_license_to_minio(self, file: UploadFile, model_id: UUID) -> str:
+    async def _save_license_file_to_minio(self, file: UploadFile, model_id: UUID) -> str:
         """Save uploaded file and return file path."""
         # Check if the license file already exists in minio
         minio_store = ModelStore()
-        license_object_prefix = f"{model_id}"
+        license_object_prefix = f"{MINIO_LICENSE_OBJECT_NAME}/{model_id}"
         try:
             # is_minio_object_exists = minio_store.check_file_exists(app_settings.minio_model_bucket, license_object_prefix)
             minio_store.remove_objects(app_settings.minio_model_bucket, license_object_prefix, recursive=True)
@@ -2120,7 +2096,7 @@ class ModelService(SessionMixin):
                 file_path = os.path.join(temp_dir, file.filename)
                 with open(file_path, "wb") as f:
                     f.write(await file.read())
-                license_object_name = f"{model_id}/{file.filename}"
+                license_object_name = f"{MINIO_LICENSE_OBJECT_NAME}/{model_id}/{file.filename}"
 
                 # Upload file to minio
                 minio_store.upload_file(app_settings.minio_model_bucket, file_path, license_object_name)
@@ -2132,21 +2108,49 @@ class ModelService(SessionMixin):
                 status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, message="Unable to save license file"
             ) from e
 
+    async def _save_license_url_to_minio(self, license_url: str, filename: str, model_id: UUID) -> str:
+        """Save uploaded file and return file path."""
+        # Check if the license file already exists in minio
+        minio_store = ModelStore()
+        license_object_prefix = f"{MINIO_LICENSE_OBJECT_NAME}/{model_id}"
+        try:
+            # is_minio_object_exists = minio_store.check_file_exists(app_settings.minio_model_bucket, license_object_prefix)
+            minio_store.remove_objects(app_settings.minio_model_bucket, license_object_prefix, recursive=True)
+
+            # Download file to a temp folder and save it to minio
+            with tempfile.TemporaryDirectory() as temp_dir:
+                response = requests.get(license_url, timeout=30)
+                response.raise_for_status()
+
+                # Save file to temp folder
+                file_path = os.path.join(temp_dir, filename)
+                with open(file_path, "wb") as f:
+                    f.write(response.content)
+
+                # Generate license object name
+                license_object_name = f"{MINIO_LICENSE_OBJECT_NAME}/{model_id}/{filename}"
+
+                # Upload file to minio
+                minio_store.upload_file(app_settings.minio_model_bucket, file_path, license_object_name)
+
+            return license_object_name
+        except (MinioException, Exception) as e:
+            logger.exception(f"Error uploading file to minio: {e}")
+            raise ClientException(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, message="Unable to save license file"
+            ) from e
+
     async def _create_or_update_license_entry(
         self,
         model_id: UUID,
         filename: str,
         license_url: str,
         current_user_id: UUID,
-        data_type: ModelLicenseObjectTypeEnum,
     ) -> None:
         """Create or update a license entry in the database."""
         # Set license source
-        if data_type == ModelLicenseObjectTypeEnum.MINIO:
-            minio_store = ModelStore()
-            license_source = minio_store.get_object_url(app_settings.minio_model_bucket, license_url)
-        else:
-            license_source = license_url
+        minio_store = ModelStore()
+        license_source = minio_store.get_object_url(app_settings.minio_model_bucket, license_url)
 
         # Check if a license entry with the given model_id exists
         existing_license = await ModelLicensesDataManager(self.session).retrieve_by_fields(
@@ -2159,7 +2163,6 @@ class ModelService(SessionMixin):
             update_license_data = {
                 "name": filename,
                 "url": license_url,
-                "data_type": data_type,
                 "faqs": None,
                 "description": None,
                 "suitability": None,
@@ -2177,7 +2180,6 @@ class ModelService(SessionMixin):
                 id=uuid4(),
                 name=filename,
                 url=license_url,
-                data_type=data_type,
                 model_id=model_id,
                 faqs=None,
                 description=None,
@@ -2327,11 +2329,12 @@ class ModelService(SessionMixin):
             logger.debug(f"Model deletion successful for {db_model.local_path}")
 
         # Delete existing license from minio
-        if db_model.model_licenses:
+        if db_model.model_licenses and not db_model.model_licenses.url.startswith(COMMON_LICENSE_MINIO_OBJECT_NAME):
             logger.debug(f"Deleting license from minio for model {db_model.id}")
-            if db_model.model_licenses.data_type == ModelLicenseObjectTypeEnum.MINIO:
-                minio_store = ModelStore()
-                minio_store.remove_objects(app_settings.minio_model_bucket, f"{db_model.id}", recursive=True)
+            minio_store = ModelStore()
+            minio_store.remove_objects(
+                app_settings.minio_model_bucket, f"{MINIO_LICENSE_OBJECT_NAME}/{db_model.id}", recursive=True
+            )
 
             # Delete license from db
             await ModelLicensesDataManager(self.session).delete_by_fields(ModelLicenses, {"model_id": db_model.id})
